fix: merge from/to proto/pojo + solve comments on the PR
