+++ isthmus/src/main/java/io/substrait/isthmus/SqlToSubstrait.java
@@ -1,27 +1,18 @@
 package io.substrait.isthmus;
 
 import com.google.common.annotations.VisibleForTesting;
-import io.substrait.extendedexpression.ExtendedExpressionProtoConverter;
-import io.substrait.extendedexpression.ImmutableExpressionReference;
-import io.substrait.extendedexpression.ImmutableExpressionType;
-import io.substrait.extendedexpression.ImmutableExtendedExpression;
 import io.substrait.extension.ExtensionCollector;
-import io.substrait.isthmus.expression.RexExpressionConverter;
-import io.substrait.isthmus.expression.ScalarFunctionConverter;
-import io.substrait.proto.ExtendedExpression;
 import io.substrait.proto.Plan;
 import io.substrait.proto.PlanRel;
 import io.substrait.relation.RelProtoConverter;
 import io.substrait.type.NamedStruct;
-import java.util.*;
+import java.util.List;
 import java.util.function.Function;
 import org.apache.calcite.plan.hep.HepPlanner;
 import org.apache.calcite.plan.hep.HepProgram;
 import org.apache.calcite.prepare.CalciteCatalogReader;
 import org.apache.calcite.rel.RelRoot;
-import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.rel.type.RelDataTypeFactory;
-import org.apache.calcite.rex.RexNode;
 import org.apache.calcite.schema.Schema;
 import org.apache.calcite.sql.SqlNode;
 import org.apache.calcite.sql.parser.SqlParseException;
@@ -32,12 +23,6 @@ import org.apache.calcite.sql2rel.StandardConvertletTable;
 
 /** Take a SQL statement and a set of table definitions and return a substrait plan. */
 public class SqlToSubstrait extends SqlConverterBase {
-
-  private final ScalarFunctionConverter functionConverter =
-      new ScalarFunctionConverter(EXTENSION_COLLECTION.scalarFunctions(), factory);
-
-  private final RexExpressionConverter rexExpressionConverter =
-      new RexExpressionConverter(functionConverter);
 
   public SqlToSubstrait() {
     this(null);
@@ -61,26 +46,6 @@ public class SqlToSubstrait extends SqlConverterBase {
   public Plan execute(String sql, String name, Schema schema) throws SqlParseException {
     var pair = registerSchema(name, schema);
     return executeInner(sql, factory, pair.left, pair.right);
-  }
-
-  /**
-   * Process to execute an SQL Expression to convert into an Extended expression protobuf message
-   *
-   * @param sqlExpression expression defined by the user
-   * @param tables of names of table needed to consider to load into memory for catalog, schema,
-   *     validate and parse sql
-   * @return extended expression protobuf message
-   * @throws SqlParseException
-   */
-  public ExtendedExpression executeSQLExpression(String sqlExpression, List<String> tables)
-      throws SqlParseException {
-    var result = registerCreateTablesForExtendedExpression(tables);
-    return executeInnerSQLExpression(
-        sqlExpression,
-        result.validator(),
-        result.catalogReader(),
-        result.nameToTypeMap(),
-        result.nameToNodeMap());
   }
 
   // Package protected for testing
@@ -126,38 +91,6 @@ public class SqlToSubstrait extends SqlConverterBase {
     return plan.build();
   }
 
-  private ExtendedExpression executeInnerSQLExpression(
-      String sqlExpression,
-      SqlValidator validator,
-      CalciteCatalogReader catalogReader,
-      Map<String, RelDataType> nameToTypeMap,
-      Map<String, RexNode> nameToNodeMap)
-      throws SqlParseException {
-    RexNode rexNode =
-        sqlToRexNode(sqlExpression, validator, catalogReader, nameToTypeMap, nameToNodeMap);
-    io.substrait.expression.Expression.ScalarFunctionInvocation func =
-        (io.substrait.expression.Expression.ScalarFunctionInvocation)
-            rexNode.accept(rexExpressionConverter);
-    NamedStruct namedStruct = TypeConverter.DEFAULT.toNamedStruct(nameToTypeMap);
-
-    ImmutableExpressionReference expressionReference =
-        ImmutableExpressionReference.builder()
-            .expressionType(ImmutableExpressionType.builder().expression(func).build())
-            .addOutputNames("new-column")
-            .build();
-
-    List<io.substrait.extendedexpression.ExtendedExpression.ExpressionReference>
-        expressionReferences = new ArrayList<>();
-    expressionReferences.add(expressionReference);
-
-    ImmutableExtendedExpression.Builder extendedExpression =
-        ImmutableExtendedExpression.builder()
-            .referredExpressions(expressionReferences)
-            .baseSchema(namedStruct);
-
-    return new ExtendedExpressionProtoConverter().toProto(extendedExpression.build());
-  }
-
   private List<RelRoot> sqlToRelNode(
       String sql, SqlValidator validator, CalciteCatalogReader catalogReader)
       throws SqlParseException {
@@ -172,20 +105,6 @@ public class SqlToSubstrait extends SqlConverterBase {
             .map(parsed -> getBestExpRelRoot(converter, parsed))
             .collect(java.util.stream.Collectors.toList());
     return roots;
-  }
-
-  private RexNode sqlToRexNode(
-      String sql,
-      SqlValidator validator,
-      CalciteCatalogReader catalogReader,
-      Map<String, RelDataType> nameToTypeMap,
-      Map<String, RexNode> nameToNodeMap)
-      throws SqlParseException {
-    SqlParser parser = SqlParser.create(sql, parserConfig);
-    SqlNode sqlNode = parser.parseExpression();
-    SqlNode validSQLNode = validator.validateParameterizedExpression(sqlNode, nameToTypeMap);
-    SqlToRelConverter converter = createSqlToRelConverter(validator, catalogReader);
-    return converter.convertExpression(validSQLNode, nameToNodeMap);
   }
 
   @VisibleForTesting
